Support IAudioExtension AIDL v4 and v5
We unfroze the v4 IAudioExtension AIDL API and modified it in
ag/30424231 and incremented the version to v5 in ag/30702418 to fix the
tests. This broke the -next builds, which expected the previously frozen
version (v4).

This change modifies the compatibility matrix to support both v4 and v5.
This fixes both the local tests and -next builds.

Bug: 378303588
Bug: 383213634
Test: atest vts_treble_vintf_vendor_test
Flag: vendor.google.whitechapel.audio.flags.enable_always_available_display
Change-Id: I467434ccb41ab71dc1b230e64dc09fc0959a8b8a